+++ GHEtool/VariableClasses/LoadData/GeothermalLoad/HourlyGeothermalLoadMultiYear.py
@@ -4,7 +4,6 @@ import warnings
 
 import numpy as np
 
-from GHEtool.logger import ghe_logger
 from GHEtool.VariableClasses.LoadData.Baseclasses import _HourlyData
 from GHEtool.VariableClasses.LoadData.GeothermalLoad.HourlyGeothermalLoad import HourlyGeothermalLoad