+++ GHEtool/VariableClasses/PipeData/_PipeData.py
@@ -3,7 +3,7 @@ import pygfunction as gt
 from abc import ABC
 from GHEtool.VariableClasses.BaseClass import BaseClass
 from GHEtool.VariableClasses.FluidData import _FluidData
-from GHEtool.VariableClasses.FlowRateData import _FlowRateData
+from GHEtool.VariableClasses.FlowRateData import _FlowData
 
 
 class _PipeData(BaseClass, ABC):
@@ -33,7 +33,7 @@ class _PipeData(BaseClass, ABC):
         self.epsilon = epsilon  # pipe roughness m
 
     @abc.abstractmethod
-    def calculate_resistances(self, fluid_data: _FluidData, flow_rate_data: _FlowRateData, **kwargs) -> None:
+    def calculate_resistances(self, fluid_data: _FluidData, flow_rate_data: _FlowData, **kwargs) -> None:
         """
         This function calculates the conductive and convective resistances, which are constant.
 
@@ -67,7 +67,7 @@ class _PipeData(BaseClass, ABC):
         """
 
     @abc.abstractmethod
-    def Re(self, fluid_data: _FluidData, flow_rate_data: _FlowRateData, **kwargs) -> float:
+    def Re(self, fluid_data: _FluidData, flow_rate_data: _FlowData, **kwargs) -> float:
         """
         Reynolds number.
 
@@ -84,7 +84,7 @@ class _PipeData(BaseClass, ABC):
         """
 
     @abc.abstractmethod
-    def pressure_drop(self, fluid_data: _FluidData, flow_rate_data: _FlowRateData, borehole_length: float,
+    def pressure_drop(self, fluid_data: _FluidData, flow_rate_data: _FlowData, borehole_length: float,
                       **kwargs) -> float:
         """
         Calculates the pressure drop across the entire borehole.